move processRequests() back to specific components
