+++ src/app/modules/group/http-services/request-actions.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { ActionResponse, successData, objectToMap } from 'src/app/shared/http-services/action-response';
 import { map } from 'rxjs/operators';
-import { Observable } from 'rxjs';
+import { forkJoin, Observable } from 'rxjs';
 import { appConfig } from 'src/app/shared/helpers/config';
 import { Result } from '../components/pending-request/pending-request-response-handling';
 
@@ -18,42 +18,54 @@ export class RequestActionsService {
 
   constructor(private http: HttpClient) {}
 
-  processJoinRequest(groupId: string, memberIds: string[], action: Action): Observable<Map<string, any>> {
+  processJoinRequests(ids: Map<string, string[]>, action: Action): Observable<Map<string, any>[]> {
     const type = action === Action.Accept ? 'accept' : 'reject';
-    return this.http
-      .post<ActionResponse<Object>>(`${appConfig().apiUrl}/groups/${groupId}/join-requests/${type}`, null, {
-        params: {
-          group_ids: memberIds.join(','),
-        },
-      })
-      .pipe(
-        map(successData),
-        map(objectToMap)
-      );
+    return forkJoin(
+      Array.from(ids.entries()).map(groupMembersIds =>
+        this.http
+          .post<ActionResponse<Object>>(`${appConfig().apiUrl}/groups/${groupMembersIds[0]}/join-requests/${type}`, null, {
+            params: {
+              group_ids: groupMembersIds[1].join(','),
+            },
+          })
+          .pipe(
+            map(successData),
+            map(objectToMap)
+          )
+      )
+    );
   }
 
-  processLeaveRequest(groupId: string, memberIds: string[], action: Action): Observable<Map<string, any>> {
+  processLeaveRequests(ids: Map<string, string[]>, action: Action): Observable<Map<string, any>[]> {
     const type = action === Action.Accept ? 'accept' : 'reject';
-    return this.http
-      .post<ActionResponse<Object>>(`${appConfig().apiUrl}/groups/${groupId}/leave-requests/${type}`, null, {
-        params: {
-          group_ids: memberIds.join(','),
-        },
-      })
-      .pipe(
-        map(successData),
-        map(objectToMap)
-      );
+    return forkJoin(
+      Array.from(ids.entries()).map(groupMembersIds =>
+        this.http
+          .post<ActionResponse<Object>>(`${appConfig().apiUrl}/groups/${groupMembersIds[0]}/leave-requests/${type}`, null, {
+            params: {
+              group_ids: groupMembersIds[1].join(','),
+            },
+          })
+          .pipe(
+            map(successData),
+            map(objectToMap)
+          )
+      )
+    );
   }
 
-  processGroupInvitation(groupId: string, action: Action): Observable<Map<string, any>> {
+  processGroupInvitations(groupIds: string[], action: Action): Observable<Map<string, any>[]> {
     const type = action === Action.Accept ? 'accept' : 'reject';
-    return this.http
-      .post<ActionResponse<Object>>(`${appConfig().apiUrl}/current-user/group-invitations/${groupId}/${type}`, null)
-      .pipe(
-        map(successData),
-        map(objectToMap)
-      );
+    return forkJoin(
+      groupIds.map(groupId =>
+        this.http
+          .post<ActionResponse<Object>>(`${appConfig().apiUrl}/current-user/group-invitations/${groupId}/${type}`, null)
+          .pipe(
+            map(successData),
+            map(objectToMap)
+          )
+      )
+    );
   }
 }
